% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survplot_functions.R
\name{calc_KMcurve_pvalues}
\alias{calc_KMcurve_pvalues}
\title{Calculate the p-value for differnces in KM curves from survdiff() function}
\usage{
calc_KMcurve_pvalues(survdiff, digits = 3)
}
\arguments{
\item{survdiff}{the results from survival::survdiff()}

\item{digits}{a numeric value of the number of digits for the p-value}
}
\value{

}
\description{
Calculate the p-value for differnces in KM curves from survdiff() function
}
\examples{
diff <- survival::survdiff(survival::Surv(time, status) ~ x, data = survival::aml)
calc_KMcurve_pvalues(diff)
}
