% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survplot_functions.R
\name{coxSummaryTable}
\alias{coxSummaryTable}
\title{Create a dataframe output to summarize the results of Cox proportional hazards regression}
\usage{
coxSummaryTable(coxph.res, Colname = "Group")
}
\arguments{
\item{coxph.res}{the results of survival::coxph()}

\item{Colname}{a character string of the column name}
}
\value{

}
\description{
Create a dataframe output to summarize the results of Cox proportional hazards regression
}
\examples{
coxreg <- survival::coxph(survival::Surv(time, status) ~ x, data = survival::aml)
coxSummaryTable(coxreg)
}
