% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_annotation_functions.R
\name{define_exon_junctions}
\alias{define_exon_junctions}
\title{Provides Exon Number from Genomic Coordinates.}
\usage{
define_exon_junctions(df, exon_ranges_GR, intron_ranges_GR = NULL)
}
\arguments{
\item{df}{dataframe for a single fusion  with the breakpoint of the fusion-genes divided into 4 columns, chromosome and position for each gene in the fusion.}

\item{exon_ranges_GR}{GRanges list object that contains exon ranges of transcripts involved in the fusions in the input dataframe. No duplicated gene IDs - one transcript ID per one gene ID}

\item{intron_ranges_GR}{GRanges list object that contains intron ranges of transcripts involved in the fusions in the input dataframe. No duplicated gene IDs - one transcript ID per one gene ID}
}
\value{

}
\description{
Provides Exon Number from Genomic Coordinates.
}
\examples{
\dontrun{
my_transcript_IDs <- c("ENSTXXX","ENSTXXXXX")
intronGR <- intronsByTranscript(Grch37.txdb, use.names=TRUE)[my_transcript_IDs]
update.intronGR <- mcol_intron_rank(intronGR)
exonGR <- exonsBy(Grch37.txdb, by="tx", use.names=TRUE)[my_transcript_IDs]
fusions.df <- data.frame(chromosome_A="chr16",pos_A=00000100, chromosome_B="chr16", pos_B=00000200)
junctions.df <- define_exon_junctions(fusions.df,exonGR,update.intronGR)
}

}
