% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_annotation_functions.R
\name{fusion_category}
\alias{fusion_category}
\title{Define Fusion Category for Harmonization Steps}
\usage{
fusion_category(geneA, geneB)
}
\arguments{
\item{geneA}{column name which contains the gene symbol for fusion partner A}

\item{geneB}{column name which contains the gene symbol for fusion partner A}
}
\value{

}
\description{
Define Fusion Category for Harmonization Steps
}
\examples{
\dontrun{
fusion.df <- data.frame(geneNameA="NUP98", geneNameB="KDM5A", Breakpoint="chr11:0000000|chr12:0000000") \%>\%
 rowwise() \%>\%
 mutate(Fusion.Category=fusion_category(geneNameA, geneNameB)) \%>\%
 ungroup()
}
}
