% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_searches_workflow.R
\name{genome_paint_format}
\alias{genome_paint_format}
\title{A function to reformat a breakpoint information for genome paint}
\usage{
genome_paint_format(df, fusion_genes_GR)
}
\arguments{
\item{df}{data frame with the column names "chomosomes_A", "pos_A", "chromosomes_B", "pos_B"}

\item{fusion_genes_GR}{the GR ranges object with the fusion genes positions}
}
\value{

}
\description{
A function to reformat a breakpoint information for genome paint
}
