% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_searches_workflow.R
\name{match_reads_contigs}
\alias{match_reads_contigs}
\title{Sequence pattern matching of RNAseq reads and counts the number of breakpoint junction reads.}
\usage{
match_reads_contigs(Sample_ID, pdict_obj, RNAseqReads, contigSeqs = NULL)
}
\arguments{
\item{Sample_ID}{character string of patient ID}

\item{pdict_obj}{pdict class object form biostrings.}

\item{RNAseqReads}{DNAstring set object with RNAseq reads to query}

\item{contigSeqs}{DNAstring set object with fusion algorithm (eg STAR-fusion) derived fusion contig sequences}
}
\value{

}
\description{
Sequence pattern matching of RNAseq reads and counts the number of breakpoint junction reads.
}
\examples{
\dontrun{
my_seqs <- data.frame(Name=paste0("seq",1:10),Sequence=rep(c("ATCGCCCGTTA"), 10))
my_files_df <- data.frame(Sample="PAZZUM.03A.01R", filepath=dir("/path/to/sample.bam"))

my_contigs <- data.frame(contig="AATGGCTATC", Name=paste0("seq",1)) \%>\% pull(contig,name=Name)
my_contigs <- Biostrings::DNAStringSet(my_contigs)

my_pdict_obj <- create_custom_pdict(theoretical_seqs_df=my_seqs)
my_bam_reads <- subset_bam("PAZZUM.03A.01R", breakpoint=list("16:1250000|16:1260000"), my_files_df)

evidence <- match_reads_contigs("PAZZUM.03A.01R", my_pdict_obj, my_bam_reads, contigSeqs=my_contigs)
}
}
