% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_annotation_functions.R
\name{mcol_intron_rank}
\alias{mcol_intron_rank}
\title{Add Intron Rank Column to GRanges containing gene introns}
\usage{
mcol_intron_rank(intron_GR)
}
\arguments{
\item{intron_GR}{GRanges list object that contains intron ranges of genes involved in the fusions in the input dataframe}
}
\value{

}
\description{
Add Intron Rank Column to GRanges containing gene introns
}
\examples{
\dontrun{
intronGR <- intronsByTranscript(Grch37.txdb, use.names=TRUE)
update.intronGR <- mcol_intron_rank(intronGR)
}
}
