% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_searches_workflow.R
\name{order_breakpoints}
\alias{order_breakpoints}
\title{Function to sort breakpoints for easier matching of identical breakpoints across different datasets}
\usage{
order_breakpoints(brkpt, sep = "|")
}
\arguments{
\item{brkpt}{a column name for the breakpoint information}

\item{sep}{is a string for the seperator used between breakpoint junctions eg, "16:15000000|16:16000000" the sep is "|" pipe.}
}
\value{

}
\description{
Function to sort breakpoints for easier matching of identical breakpoints across different datasets
}
\examples{
\dontrun{
my_fusions <- fusion_df \%>\%
rowwise() \%>\%
mutate(Breakpoint_Ordered=order_breakpoints(breakpoint_column)) \%>\%
ungroup()
}
}
