% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survplot_functions.R
\name{outcome_table}
\alias{outcome_table}
\title{Create a table with OS/EFS percent at specified time point}
\usage{
outcome_table(fit, time = 5, pvalues = NULL)
}
\arguments{
\item{fit}{results of survival::survfit()}

\item{time}{the time point at which to summarize OS \% and EFS \%}

\item{pvalues}{a string of the p-value to include in the table}
}
\value{

}
\description{
Create a table with OS/EFS percent at specified time point
}
\examples{
fit <-  survival::survfit(survival::Surv(time, status) ~ x, data = survival::aml)
outcome_table(fit)
}
