% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survplot_functions.R
\name{risk.table}
\alias{risk.table}
\title{Title}
\usage{
risk.table(
  survFit.obj,
  f.levels = NULL,
  times = NULL,
  col = "black",
  ret.data = FALSE,
  max.year = NULL
)
}
\arguments{
\item{survFit.obj}{the resulting object from  survival::survfit()}

\item{f.levels}{factor levels of the groups}

\item{times}{a numeric vector of the summary times for the KM estimates}

\item{col}{a string indicating the text color}

\item{ret.data}{boolean - return the summary table or not}

\item{max.year}{a numeric of the the x-axis limit}
}
\value{

}
\description{
Title
}
\examples{
fit <-  survival::survfit(survival::Surv(time, status) ~ x, data = survival::aml)
risk.table(survFit.obj=fit)
}
