% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_searches_workflow.R
\name{subset_bam}
\alias{subset_bam}
\title{Subset a bam for specific chromome(s) or ranges based on input breakpoint or fusion data file.}
\usage{
subset_bam(
  Sample_ID,
  breakpoint = NULL,
  fusion_data = NULL,
  file_manifest,
  by_chr = FALSE,
  scan = FALSE,
  Seqlens = NULL,
  outdir = file.path("/fh/scratch/delete90/jlsmith3/subset_bams")
)
}
\arguments{
\item{Sample_ID}{character string of patient ID}

\item{breakpoint}{character vector with breakpoint string (can be named)}

\item{fusion_data}{dataframe with fusion results from CICERO, TransAbyss, and STAR-Fusion concatenated}

\item{file_manifest}{dataframe with Sample_ID and the full file path of the BAM file.}

\item{by_chr}{boolean - get reads mapped to the entire chromosome containing the breakpoint or just +/- 1e6 bp from breakpoint}

\item{scan}{boolean - read into memory or not}

\item{Seqlens}{named numeric vector with chromosome lengths}

\item{outdir}{destination for subsetted bam files if scan = FALSE}
}
\value{

}
\description{
Subset a bam for specific chromome(s) or ranges based on input breakpoint or fusion data file.
}
\examples{
\dontrun{
my_files_df <- data.frame(Sample="PAZZUM.03A.01R", filepath=dir("/path/to/sample.bam"))
subset_bam("PAZZUM.03A.01R", breakpoint=list("16:1250000|16:1260000"), my_files_df)
}
}
