% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survplot_functions.R
\name{SurvivalPlot}
\alias{SurvivalPlot}
\title{Create Kaplan-Meier plots}
\usage{
SurvivalPlot(fit, LegendTitle, timeUnit, colors, pval = NULL, max.year = NULL)
}
\arguments{
\item{fit}{the resulting object from surival::survfit}

\item{LegendTitle}{a string}

\item{timeUnit}{a string (years, months, days, etc)}

\item{colors}{a character vector of the same length as the number of groups in the KM plot}

\item{pval}{an option string wit the pvalue}

\item{max.year}{a numeric value to set x-axis limits.}
}
\value{

}
\description{
Create Kaplan-Meier plots
}
\examples{
fit <-  survival::survfit(survival::Surv(time, status) ~ x, data = survival::aml)
SurvivalPlot(fit=fit, LegendTitle="AML", timeUnit="months", colors=c("red","blue"))
}
