% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_searches_workflow.R
\name{create_custom_pdict}
\alias{create_custom_pdict}
\title{Title}
\usage{
create_custom_pdict(theoretical_seqs_df)
}
\arguments{
\item{theoretical_seqs_df}{a dataframe with computationally derived exon junction sequences}
}
\value{

}
\description{
Title
}
\examples{
my_seqs <- data.frame(Name=paste0("seq",1:10),Sequence=rep(c("ATCGCCCGTTA"), 10))
my_pdict_obj <- create_custom_pdict(theoretical_seqs_df=my_seqs)

}
