% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_searches_workflow.R
\name{define_exon_junctions}
\alias{define_exon_junctions}
\title{Function to define the exon number for each gene partner in fusion.}
\usage{
define_exon_junctions(df, exon_ranges_GR, intron_ranges_GR = NULL)
}
\arguments{
\item{df}{data frame with the column names "Sample","chomosomes_A", "pos_A", "chromosomes_B", "pos_B"}

\item{exon_ranges_GR}{the GRanges object with the fusion genes exon positions. One transcript ID per gene symbol.}

\item{intron_ranges_GR}{the GRanges object with the fusion genes intron positions.One transcript ID per gene symbol. Must have mcol of "intron_rank".}
}
\value{

}
\description{
Function to define the exon number for each gene partner in fusion.
}
