% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_searches_workflow.R
\name{examine_breakpoint_evidence}
\alias{examine_breakpoint_evidence}
\title{Title}
\usage{
examine_breakpoint_evidence(
  Sample_ID,
  fusion_data = NULL,
  breakpoint = NULL,
  file_manifest,
  Seqlens,
  fusion_theoretical_pdict,
  cicero.contig,
  TA.contigs
)
}
\arguments{
\item{Sample_ID}{character string of patient ID}

\item{fusion_data}{dataframe with fusion results from CICERO, TransAbyss, and STAR-Fusion concatenated}

\item{breakpoint}{list class with breakpoint string}

\item{file_manifest}{dataframe with Sample_ID and the full file path of the BAM file.}

\item{Seqlens}{named numeric vector with chromosome lengths}

\item{fusion_theoretical_pdict}{a pdict obj with computationally derived exon junction sequences}

\item{cicero.contig}{a dataframe with contig sequences from CICERO output}

\item{TA.contigs}{a dataframe with contig sequences from TransAbyss output}
}
\value{

}
\description{
Title
}
\examples{
examine_breakpoint_evidence("PAZZUM.03A.01R", breakpoint=list("16:1250000|16:1260000"), my_bam_manifest,
chr_lengths, my_pdict_obj, cicero_contig_df, transabyss_contig_df)

}
