% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_searches_workflow.R
\name{fusion_category}
\alias{fusion_category}
\title{Function to create a fusion category by alphabetical order}
\usage{
fusion_category(geneA, geneB)
}
\arguments{
\item{geneA}{the column name that contains fusion partner #1}

\item{geneB}{the column name that contains fusion partner #2}
}
\value{

}
\description{
Function to create a fusion category by alphabetical order
}
\examples{
\dontrun{
my_fusions <- fusion_df \%>\%
rowwise() \%>\%
mutate(Fusion.Category=fusion_category(geneA=geneNameA, geneB=geneNameB)) \%>\%
ungroup()
}
}
