% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_annotation_functions.R
\name{genome_paint_format}
\alias{genome_paint_format}
\title{Create JSON file for input into St. Jude Genome Paint}
\usage{
genome_paint_format(df, fusion_genes_GR)
}
\arguments{
\item{df}{dataframe for a single fusion  with the breakpoint of the fusion-genes divided into 4 columns, chromosome and position for each gene in the fusion.}

\item{fusion_genes_GR}{fusion_genes_GR is granges list object that contains genes that involved in the fusions in the input dataframe}
}
\value{

}
\description{
Create JSON file for input into St. Jude Genome Paint
}
\examples{
\dontrun{
genesGR <- transcriptsBy(Grch37.txdb, by="gene")
fusion.df <- data.frame(Fusion.Category="CBFB-MYH11",
 FusionID="My Fusion Name",
 chromosomes_A="chr16",
 pos_A=00000100,
 chromosomes_B="chr16",
 pos_B=00000200)

formatted <- genome_paint_format(df=fusion.df, fusion_genes_GR = genesGR)

}
}
