% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_searches_workflow.R
\name{match_reads_contigs}
\alias{match_reads_contigs}
\title{Title}
\usage{
match_reads_contigs(Sample_ID, pdict_obj, RNAseqReads, contigSeqs = NULL)
}
\arguments{
\item{Sample_ID}{character string of patient ID}

\item{pdict_obj}{pdict class object form biostrings.}

\item{RNAseqReads}{DNAstring set object with RNAseq reads to query}

\item{contigSeqs}{DNAstring set object with fusion algorithm (eg STAR-fusion) derived fusion contig sequences}
}
\value{

}
\description{
Title
}
\examples{
match_reads_contigs("PAZZUM.03A.01R", my_pdict_obj, my_bam_reads, contigSeqs=my_contigs)
}
