% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_annotation_functions.R
\name{order_breakpoints}
\alias{order_breakpoints}
\title{Order breakpoints Numerically for harmonization steps}
\usage{
order_breakpoints(brkpt, sep = "|")
}
\arguments{
\item{brkpt}{the column name containing the breakpoint position}

\item{sep}{the seperator between the two breakpoints}
}
\value{

}
\description{
Order breakpoints Numerically for harmonization steps
}
\examples{
\dontrun{
fusion.df <- data.frame(geneNameA="NUP98", geneNameB="KDM5A", Breakpoint="chr12:0000000|chr11:0000000") \%>\%
 rowwise() \%>\%
 mutate(Breakpoint.Ordered=order_breakpoints(Breakpoint)) \%>\%
 ungroup()

}
}
